horiz and vert flipping for HText

+++ java/src/hype/drawable/HImage.java
@@ -95,8 +95,8 @@ public class HImage extends HDrawable {
 		currAlphaPerc *= (_fill>>>24);
 		g.tint( _fill | 0xFF000000, H.app().round(currAlphaPerc) );
 		
-		float wscale;
-		float hscale;
+		int wscale;
+		int hscale;
 		float w;
 		float h;
 		if(_width < 0) {
@@ -117,8 +117,8 @@ public class HImage extends HDrawable {
 		}
 		
 		g.pushMatrix();
-		g.scale(wscale, hscale);
-		g.image(_image, drawX,drawY, w,h);
+			g.scale(wscale, hscale);
+			g.image(_image, drawX,drawY, w,h);
 		g.popMatrix();
 	}
 }
